Extract `getApi`.
