+++ frontend/src/app/functions/index.ts
@@ -1,1 +1,2 @@
+export { getApi } from './getApi';
 export { processMessages } from './processMessages';
